Refactor/#147 modal component

+++ src/features/resume/draft-resumes-modal.tsx
@@ -3,10 +3,10 @@
 import { Dispatch, SetStateAction } from 'react';
 import Modal from '@/components/ui/modal';
 import Typography from '@/components/ui/typography';
-import type { Resume } from '@prisma/client';
 import { MODAL_ID } from '@/constants/modal-id-constants';
 import { useDeleteResumeMutation } from '@/features/resume/hooks/use-delete-resume-mutation';
-import Trash from '@/components/icons/trash';
+import DraftResumeItem from '@/features/resume/draft-resume-item';
+import type { Resume } from '@prisma/client';
 
 const { DRAFT_RESUME } = MODAL_ID;
 
@@ -22,7 +22,10 @@ const DraftResumesModal = ({ draftResumeList, isError, onLoadDraft, activeResume
   const { mutate: deleteResumeMutate } = useDeleteResumeMutation();
 
   const handleDeleteResume = (resumeId: number) => {
-    deleteResumeMutate(resumeId);
+    if (window.confirm('자기소개서를 정말로 삭제하시겠습니까?')) {
+      deleteResumeMutate(resumeId);
+    }
+
     if (activeResumeId === resumeId) {
       setResumeId(null);
     }
@@ -33,18 +36,29 @@ const DraftResumesModal = ({ draftResumeList, isError, onLoadDraft, activeResume
   };
 
   return (
-    <Modal modalId={DRAFT_RESUME} className='flex flex-col gap-2'>
+    <Modal modalId={DRAFT_RESUME} className='flex flex-col gap-4'>
+      <div>
+        <Typography as='h2' size='2xl' weight='bold' align='center'>
+          임시 저장된 자소서
+        </Typography>
+        <Typography color='primary-600' weight='bold' align='center'>
+          작성 완료 시 300 경험치 획득!
+        </Typography>
+      </div>
       {!isError && draftResumeList?.length === 0 ? (
-        <Typography>임시 저장된 자기소개서가 없습니다</Typography>
+        <Typography color='gray-500'>임시 저장된 자기소개서가 없습니다</Typography>
       ) : (
-        draftResumeList?.map((resume) => (
-          <div key={resume.id} className='flex justify-between'>
-            <button onClick={() => handleDraftResumeClick(resume)}>{resume.title}</button>
-            <button onClick={() => handleDeleteResume(resume.id)}>
-              <Trash />
-            </button>
-          </div>
-        ))
+        <ul className='flex flex-col gap-4'>
+          {draftResumeList?.map((resume) => {
+            return (
+              <DraftResumeItem
+                resume={resume}
+                onDeleteClick={handleDeleteResume}
+                onDraftResumeClick={handleDraftResumeClick}
+              />
+            );
+          })}
+        </ul>
       )}
     </Modal>
   );
